extracted setMultiplexerConfiguratio as a Operation * Reason: Method contains too much logic and it is certainly of interest for API purposes * The Multiplexer Config helds now information on current family and model. If we have more Information, we should consider to use a "Hearing Instrument" Class. * Renaming: family and model aka current family and model to HiFamily and HiModel
