+++ UserAgent/Backend/OperationPartClasses/GetStatus.cs
@@ -62,11 +62,11 @@ namespace RaspberryBackend
             {
                 xml.Add(new XElement("Multiplexer",
                     new XElement("Initialized", true),
-                    new XElement("Family", this.Multiplexer.family),
-                    new XElement("ModelName", this.Multiplexer.model_name)
+                    new XElement("Family", MultiplexerConfig.HiFamily),
+                    new XElement("ModelName", MultiplexerConfig.HiModel)
                 ));
                 /* Read config from XML and based on that return supported controls of current HI*/
-                XPinConfig mux_config = HiXmlParser.getMultiplexerConfig(this.Multiplexer.family, this.Multiplexer.model_name);
+                XPinConfig mux_config = HiXmlParser.getMultiplexerConfig(MultiplexerConfig.HiFamily, MultiplexerConfig.HiModel);
                 xml.Add(new XElement("HI",
                     new XElement(YPinConfig.ROCKERSWITCH_STRING, string.Join(",", mux_config.ValueToPins(YPinConfig.ROCKERSWITCH_STRING))),
                     new XElement(YPinConfig.GROUND, string.Join(",", mux_config.ValueToPins(YPinConfig.GROUND))),
