+++ UserAgent/Backend/OperationPartClasses/SetHI.cs
@@ -1,7 +1,4 @@
-﻿using CommonFiles.TransferObjects;
-using System;
-
-namespace RaspberryBackend
+﻿namespace RaspberryBackend
 {
     /// <summary>
     /// This class represents a Command. It is a partial Operation class
@@ -11,11 +8,11 @@ namespace RaspberryBackend
         /// <summary>
         /// Execute the command SetHI
         /// </summary>
-        /// <param name="param">List of family and model in this order.</param>
-        /// <returns>The model when successfully updated config.</returns>
+        /// <param name="param">List of HiFamily and HiModel in this order.</param>
+        /// <returns>The HiModel when successfully updated config.</returns>
         public string SetHI(string family, string model)
         {
-            Multiplexer.setMultiplexerConfiguration(family, model);
+            setMultiplexerConfiguration(family, model);
             this.updateLCD();
             return model;
         }
